feat: Dependency-inject custom logger into sidecar
