+++ packages/sidecar/server.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 const { setupSidecar } = await import('./dist/main.js');
 const port = process.argv[2];
-setupSidecar(port);
+setupSidecar({ port });
