+++ packages/sidecar/src/main.ts
@@ -1,87 +1,28 @@
 import { createWriteStream } from 'fs';
 import { IncomingMessage, Server, ServerResponse, createServer } from 'http';
-import kleur from 'kleur';
 import { createGunzip, createInflate } from 'zlib';
+import { SidecarLogger, activateLogger, logger } from './logger.js';
+import { MessageBuffer } from './messageBuffer.js';
 
-function generateUuidv4(): string {
-  let dt = new Date().getTime();
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
-    let rnd = Math.random() * 16;
-    rnd = (dt + rnd) % 16 | 0;
-    dt = Math.floor(dt / 16);
-    return (c === 'x' ? rnd : (rnd & 0x3) | 0x8).toString(16);
-  });
-}
-
-class MessageBuffer<T> {
-  private size: number;
-  private items: [number, T][];
-  private writePos = 0;
-  private head = 0;
-  private timeout = 10;
-  private readers = new Map<string, (item: T) => void>();
-  private timeoutId: NodeJS.Timeout | undefined;
-
-  constructor(size = 100) {
-    this.size = size;
-    this.items = new Array(size);
-  }
-
-  put(item: T): void {
-    const curTime = new Date().getTime();
-    this.items[this.writePos % this.size] = [curTime, item];
-    this.writePos += 1;
-    if (this.head === this.writePos) {
-      this.head += 1;
-    }
-
-    const minTime = curTime - this.timeout * 1000;
-    let atItem;
-    while (this.head < this.writePos) {
-      atItem = this.items[this.head % this.size];
-      if (atItem === undefined) break;
-      if (atItem[0] > minTime) break;
-      this.head += 1;
-    }
-  }
-
-  subscribe(callback: (item: T) => void): string {
-    const readerId = generateUuidv4();
-    this.readers.set(readerId, callback);
-    clearTimeout(this.timeoutId);
-    this.timeoutId = setTimeout(() => this.stream(readerId));
-    return readerId;
-  }
-
-  unsubscribe(readerId: string): void {
-    this.readers.delete(readerId);
-    clearTimeout(this.timeoutId);
-  }
-
-  stream(readerId: string, readPos = this.head): void {
-    clearTimeout(this.timeoutId);
-
-    const cb = this.readers.get(readerId);
-    if (!cb) return;
-
-    let atReadPos = readPos;
-    let item;
-    /* eslint-disable no-constant-condition */
-    while (true) {
-      item = this.items[atReadPos % this.size];
-      // atReadPos >= this.writePos prevents the case where we have a full buffer
-      if (typeof item === 'undefined' || atReadPos >= this.writePos) {
-        break;
-      }
-      cb(item[1]);
-      atReadPos += 1;
-    }
-
-    this.timeoutId = setTimeout(() => this.stream(readerId, atReadPos), 500);
-  }
-}
+const DEFAULT_PORT = 8969;
 
 type Payload = [string, string];
+
+type SideCarOptions = {
+  /**
+   * The port on which the sidecar should listen.
+   * Defaults to 8969.
+   */
+  port?: string | number;
+
+  /**
+   * A logger that implements the SidecarLogger interface.
+   * Use this to inject your custom logger implementation.
+   *
+   * @default - a simple logger logging to the console.
+   */
+  logger?: SidecarLogger;
+};
 
 function getCorsHeader(): { [name: string]: string } {
   return {
@@ -103,7 +44,7 @@ function handleStreamRequest(req: IncomingMessage, res: ServerResponse, buffer: 
       res.flushHeaders();
 
       const sub = buffer.subscribe(([payloadType, data]) => {
-        log(`🕊️ sending to Spotlight`);
+        logger.debug(`🕊️ sending to Spotlight`);
         res.write(`event:${payloadType}\n`);
         // This is very important - SSE events are delimited by two newlines
         data.split('\n').forEach(line => {
@@ -129,7 +70,7 @@ function handleStreamRequest(req: IncomingMessage, res: ServerResponse, buffer: 
         });
         res.end();
       } else if (req.method === 'POST') {
-        log(`📩 Received event`);
+        logger.debug(`📩 Received event`);
         let body: string = '';
         let stream = req;
 
@@ -160,7 +101,7 @@ function handleStreamRequest(req: IncomingMessage, res: ServerResponse, buffer: 
             const filename = `${contentType.replace(/[^a-z0-9]/gi, '_').toLowerCase()}-${timestamp}.txt`;
 
             createWriteStream(filename).write(body);
-            log(`🗃️ Saved data to ${filename}`);
+            logger.info(`🗃️ Saved data to ${filename}`);
           }
 
           res.writeHead(204, {
@@ -192,15 +133,10 @@ function startServer(buffer: MessageBuffer<Payload>, port: number): Server {
     }
   });
   server.listen(port, () => {
-    log(`Sidecar listening on ${port}`);
+    logger.info(`Sidecar listening on ${port}`);
   });
 
   return server;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function log(...args: any[]) {
-  console.log(kleur.bold(kleur.magenta('🔎 [Spotlight]')), ...args);
 }
 
 let serverInstance: Server;
@@ -213,11 +149,15 @@ const isValidPort = (value: string | number) => {
   return value > 0 && value <= 65535;
 };
 
-export function setupSidecar(port?: string | number): void {
-  let sidecarPort = 8969;
+export function setupSidecar({ port, logger: customLogger }: SideCarOptions = {}): void {
+  let sidecarPort = DEFAULT_PORT;
+
+  if (customLogger) {
+    activateLogger(customLogger);
+  }
 
   if (port && !isValidPort(port)) {
-    log('Please provide a valid port.');
+    logger.info('Please provide a valid port.');
     process.exit(1);
   } else if (port) {
     sidecarPort = typeof port === 'string' ? parseInt(port, 10) : port;
@@ -232,7 +172,7 @@ export function setupSidecar(port?: string | number): void {
 
 function shutdown() {
   if (serverInstance) {
-    log('Shutting down server');
+    logger.info('Shutting down Server');
     serverInstance.close();
   }
 }
