+++ packages/spotlight/bin/run.js
@@ -1,4 +1,4 @@
 #!/usr/bin/env node
 const { setupSidecar } = await import('../dist/sidecar.js');
 const port = process.argv[2];
-setupSidecar(port);
+setupSidecar({ port });
